controlli sulla registrazione, ordini utente, amministratore può evadere e rifiutare, bottoni vari, controlli su quantità e disponibilità e altre cose che non ricordo

+++ progettojorick/src/it/progettojorick/dao/mysql/RichiestaOrdineDAO.java
@@ -82,7 +82,7 @@ public class RichiestaOrdineDAO implements IRichiestaOrdineDAO {
             r.setStato(riga[2]);
             r.setCarrello(CarrelloDAO.getInstance().findById(Integer.parseInt(riga[3])));
             r.setUtente(UtenteDAO.getInstance().findByEmail(riga[4]));
-            r.setPagamento(PagamentoDAO.getInstance().findByNumcarta(Integer.parseInt(riga[5])));
+            r.setPagamento(PagamentoDAO.getInstance().findByNumcarta(Long.parseLong(riga[5])));
             listaRichieste.add(r);
         }
         return listaRichieste;
